Remove separate directory for bus and datastore
They are likely not going to become multifile.
The bus/bus, datastore/datastore imports was long
and felt overkill.
